gumbo: Fix MSVC warnings about conversion from 'size_t' to 'int'
Fix warnings like
  warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
  warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

with Microsoft (R) C/C++ Optimizing Compiler Version 19.29.30037 for x64

+++ src/gumbo/error.c
@@ -33,7 +33,7 @@
 static int print_message(
     GumboParser* parser, GumboStringBuffer* output, const char* format, ...) {
   va_list args;
-  int remaining_capacity = output->capacity - output->length;
+  size_t remaining_capacity = output->capacity - output->length;
   va_start(args, format);
   int bytes_written = vsnprintf(
       output->data + output->length, remaining_capacity, format, args);
